Rename `repr.rs` to `representation.rs`
